refactor code and setup linters
