+++ modules/todoArr.js
@@ -1,0 +1,19 @@
+const todoArr = [
+  {
+    index: 0,
+    completed: false,
+    discription: 'wash the dishes',
+  },
+  {
+    index: 1,
+    completed: false,
+    discription: 'check match time',
+  },
+  {
+    index: 2,
+    completed: false,
+    discription: 'go live on instagram',
+  },
+];
+
+export default todoArr;
