+++ package.json
@@ -272,9 +272,18 @@
     "yallist": "^4.0.0"
   },
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.3",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.27.5",
+    "hint": "^7.1.6",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1"
@@ -286,5 +295,13 @@
   },
   "keywords": [],
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/peter-akhigbe/todo-list-project.git"
+  },
+  "bugs": {
+    "url": "https://github.com/peter-akhigbe/todo-list-project/issues"
+  },
+  "homepage": "https://github.com/peter-akhigbe/todo-list-project#readme"
 }
